1.Entity(Account: add property roles in class Account), create a entity Role, RoleName 2. create a package jwt and service  to generate a string json 3.create package response and request like package DTO 4.Controller: run jwtAuthentication at package controller to login 5. parent pom ?? inject dependence  some library at parent pom

+++ mp3zing-service/src/main/java/codegym/mp3zingservice/service/AccountService.java
@@ -13,7 +13,7 @@ public interface AccountService {
 
     AccountDTO updatePassword(AccountDTO accountDTO);
 
-    AccountDTO findById(Integer id);
+    AccountDTO findById(Long id);
     void updateUser(AccountDTO accountDTO);
 }
